fix: extract paths building logic into separate function
